Revert "add new database/jwt/cookies for users"
This reverts commit 8f526d01ad86bac3f8ee756283be0b2ea5f50c87.
