+++ frontend/login/index.html
@@ -6,7 +6,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Login</title>
   <link rel="stylesheet" href="./login-page.css">
-
+  <script defer src="./login-page.js"></script>
 </head>
 
 <body>
@@ -17,14 +17,14 @@
       <p id="login-error-msg">Invalid username <span id="error-msg-second-line">and/or password</span></p>
     </div>
     
-    <form action="/authenticate" method="POST" id="login-form">  
+    <form target="_self"  id="login-form">  
       <input type="text" name="username" id="username-field" class="login-form-field" placeholder="Username">
       <input type="password" name="password" id="password-field" class="login-form-field" placeholder="Password">
-      <input type="submit" value="Login" id="login-form-submit">
+      <input type="submit" form=login-form" value="Login" id="login-form-submit">
     </form>
-    
-    <a id="login-register" target="_self" href="../register/index.html"> register </a>
- 
+    <button id="login-register">
+    <a target="_self" href="../register/index.html">Register</a>
+  </button>
   </main>
 </body>
