+++ package.json
@@ -19,13 +19,6 @@
   "homepage": "https://github.com/abbigrundy/nasa-project#readme",
   "dependencies": {
     "express": "^4.17.1",
-    "nodemon": "^2.0.6",
-    "sequelize": "^6.3.5",
-    "cookie-parser": "^1.4.5",
-    "jsonwebtoken": "^8.5.1",
-    "jwt-decode": "^3.1.2",
-    "passport": "^0.4.1",
-    "passport-jwt": "^4.0.0",
-    "pg": "^8.5.1"
+    "nodemon": "^2.0.6"
   }
 }
