--- server/database/.database-credentials.js
@@ -1,6 +1,0 @@
-const database_credentials = {
-  username: "postgres",
-  password: "Frankie58",
-  PRIVATEKEY: "1234"
-};
-module.exports = database_credentials;
