--- server/middleware/newUser.js
@@ -1,19 +1,0 @@
-const { Users } = require("../database/database");
-
-const newUser = async (request, response, next) => {
-  const user = await Users.create(
-    {
-      username: request.body.username,
-      password: request.body.password,
-      email: request.body.email,
-      isAdmin: true,
-    },
-    {
-      fields: ["username", "password", "email"],
-    }
-  );
-  next();
-  response.redirect("/login")
-};
-
-module.exports = { newUser };
