+++ styles/homepage.css
@@ -1,0 +1,10 @@
+.header {
+  background: teal;
+  color: rgb(255, 255, 255);
+  text-align: center;
+  padding: 50px;
+}
+
+.h1 {
+  text-align: center;
+}
