+++ styles/mars-rover-styles.css
@@ -1,0 +1,19 @@
+body {
+  margin: 0;
+  padding: 0;
+  height: 100hv;
+}
+
+#journey {
+  position: absolute;
+  width: 100%;
+  height: auto;
+  object-fit: fill;
+}
+
+#mars-sprite {
+  width: 15%;
+  position: relative;
+  left: 94px;
+  top: 520px;
+}
